+++ metadata.json
@@ -16,6 +16,7 @@
             "0.2.0": "e9d3fb762d7f63a2068da93b806ea97975765b4db06473ba78a00fe9127b5a4a"
         }
     },
+    "wit_version": 0,
     "external_url": "",
     "animation_url": ""
 }
