+++ ui/src/index.css
@@ -76,7 +76,7 @@ label {
 }
 
 button:hover {
-  background-color: #00000011;
+  background-color: #00000055;
 }
 
 textarea, input[type="text"] {
